Move folder conversation item to its own file
